backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
httpx>=0.27,<1.0
google-genai>=1.12,<2.0
python-dotenv>=1.0,<2.0
pyyaml>=6.0,<7.0
